+++ sail/Types/UserMeta.php
@@ -76,7 +76,7 @@ class UserMeta implements Castable, \JsonSerializable
      *
      * Get a dynamic property
      *
-     * @param string $name
+     * @param  string  $name
      * @return mixed
      *
      */
@@ -93,8 +93,8 @@ class UserMeta implements Castable, \JsonSerializable
      *
      * Set a dynamic property
      *
-     * @param string $name
-     * @param mixed $value
+     * @param  string  $name
+     * @param  mixed   $value
      * @return void
      */
     public function __set(string $name, mixed $value): void
@@ -116,7 +116,7 @@ class UserMeta implements Castable, \JsonSerializable
      *
      * Check if property exists with any type of value
      *
-     * @param string $name
+     * @param  string  $name
      * @return bool
      *
      */
@@ -133,9 +133,9 @@ class UserMeta implements Castable, \JsonSerializable
      *
      * Register an available meta
      *
-     * @param string $key
-     * @param int $type
-     * @param callable|null $callback
+     * @param  string         $key
+     * @param  int            $type
+     * @param  callable|null  $callback
      * @return void
      *
      */
@@ -148,7 +148,7 @@ class UserMeta implements Castable, \JsonSerializable
      *
      * Register a flag
      *
-     * @param string $key
+     * @param  string  $key
      * @return void
      *
      */
@@ -161,7 +161,7 @@ class UserMeta implements Castable, \JsonSerializable
      *
      * Get all available meta registered in the system
      *
-     * @param bool $inputs
+     * @param  bool  $inputs
      * @return string
      *
      */
@@ -190,7 +190,7 @@ class UserMeta implements Castable, \JsonSerializable
 
                 case self::TYPE_CUSTOM:
                     $input = ($inputs) ? 'Input' : '';
-                    $graphql .= Text::snakeCase($key) . ": {$key}{$input}!\n";
+                    $graphql .= Text::snakeCase($key) . ": {$key}{$input}\n";
                     break;
             }
         }
@@ -251,7 +251,7 @@ class UserMeta implements Castable, \JsonSerializable
      *
      * Cast to UserMeta
      *
-     * @param mixed $value
+     * @param  mixed  $value
      * @return UserMeta
      *
      */
